gcp_pipeline/collect/cloudfunctions/requirements.txt
functions-framework==3.*
Flask
google-cloud-storage
beautifulsoup4
bs4
geckodriver-autoinstaller
icecream
numpy
pandas
python-decouple
python-swiftclient
requests
scikit-learn
scipy
selenium
splinter
statsmodels
dpath
python-keystoneclient
openpyxl
matplotlib
seaborn
keras
tensorflow
requests_html
prettytable
